feat/19-implement-backend-for-notification-settings
