+++ src/controllers/index.ts
@@ -1,3 +1,4 @@
 // silence is golden
 export * from "./AuthController";
 export * from "./UserController";
+export * from "./NotificationController"
