+++ src/routes/index.ts
@@ -2,3 +2,4 @@
 export * from "./auth";
 export * from "./user";
 export * from "./testimonial";
+export * from "./notificationsettings"
